Formatting changes, comment cleaning

+++ gradient_boosting.py
@@ -13,27 +13,33 @@ DATA_PATH = r'data\best_picture_metadata_with_sampled_english_reviews.csv'
 
 def load_and_preprocess_data(filepath):
     df = pd.read_csv(filepath)
-    df['Review Text'] = df['Review Text'].astype(str).str.lower()
-    df['Year Nominated'] = df['Year Nominated'].astype(int)
-    df['Won'] = df['Won'].astype(int)
+    df["Review Text"] = df["Review Text"].astype(str).str.lower()
+    df["Year Nominated"] = df["Year Nominated"].astype(int)
+    df["Won"] = df["Won"].astype(int)
     return df
 
+
 def get_custom_stop_words(df):
-    stop_words = set(stopwords.words('english'))
-    for film in df['Film Name'].unique():
-        words = word_tokenize(re.sub(r'[^\w\s]', '', film.lower()))
+    stop_words = set(stopwords.words("english"))
+    for film in df["Film Name"].unique():
+        words = word_tokenize(re.sub(r"[^\w\s]", "", film.lower()))
         stop_words.update(words)
     return list(stop_words)
 
+
 def evaluate_year(test_year, df, review_vectorizer, clf):
-    train_mask = df['Year Nominated'] != test_year
-    test_mask = df['Year Nominated'] == test_year
+    train_mask = df["Year Nominated"] != test_year
+    test_mask = df["Year Nominated"] == test_year
 
-    X_train_text = review_vectorizer.fit_transform(df.loc[train_mask, 'Review Text']).toarray()
-    y_train = df.loc[train_mask, 'Won']
+    X_train_text = review_vectorizer.fit_transform(
+        df.loc[train_mask, "Review Text"]
+    ).toarray()
+    y_train = df.loc[train_mask, "Won"]
 
-    X_test_text = review_vectorizer.transform(df.loc[test_mask, 'Review Text']).toarray()
-    y_test = df.loc[test_mask, 'Won']
+    X_test_text = review_vectorizer.transform(
+        df.loc[test_mask, "Review Text"]
+    ).toarray()
+    y_test = df.loc[test_mask, "Won"]
 
     clf.fit(X_train_text, y_train)
     y_pred = clf.predict(X_test_text)
@@ -43,15 +49,19 @@ def evaluate_year(test_year, df, review_vectorizer, clf):
     y_prob_normalized = y_prob / y_prob.sum() if y_prob.sum() > 0 else y_prob
 
     test_df = df[test_mask].copy()
-    test_df['win_prob'] = y_prob_normalized
+    test_df["win_prob"] = y_prob_normalized
 
-    sorted_df = test_df.sort_values(by='win_prob', ascending=False).reset_index(drop=True)
+    sorted_df = test_df.sort_values(by="win_prob", ascending=False).reset_index(
+        drop=True
+    )
     predicted_winner = sorted_df.iloc[0]
-    actual_winner = sorted_df[sorted_df['Won'] == 1].iloc[0] if any(sorted_df['Won'] == 1) else None
+    actual_winner = (
+        sorted_df[sorted_df["Won"] == 1].iloc[0] if any(sorted_df["Won"] == 1) else None
+    )
 
     correct_winner = (actual_winner is not None and predicted_winner['Film Name'] == actual_winner['Film Name'])
 
-    print(f'\n--- Probabilities for {test_year} ---')
+    print(f"\n--- Probabilities for {test_year} ---")
     for idx, row in sorted_df.iterrows():
         print(f"{row['Film Name']}: {row['win_prob']:.4f}")
 
@@ -63,10 +73,11 @@ def evaluate_year(test_year, df, review_vectorizer, clf):
         'correct_prediction': correct_winner
     }
 
+
 def main():
     df = load_and_preprocess_data(DATA_PATH)
 
-    required_columns = {'Film Name', 'Year Nominated', 'Won', 'Review Text'}
+    required_columns = {"Film Name", "Year Nominated", "Won", "Review Text"}
     if not required_columns.issubset(df.columns):
         missing = required_columns - set(df.columns)
         raise ValueError(f"Missing required columns: {missing}")
@@ -74,9 +85,11 @@ def main():
     custom_stop_words = get_custom_stop_words(df)
 
     review_vectorizer = TfidfVectorizer(
+      
         max_features=5000,
         stop_words=custom_stop_words,
         ngram_range=(1, 2)
+
     )
 
     # Define the GradientBoostingClassifier with the best parameters
@@ -91,7 +104,7 @@ def main():
     results = []
     feature_importances = []
 
-    for test_year in sorted(df['Year Nominated'].unique()):
+    for test_year in sorted(df["Year Nominated"].unique()):
         print(f"\n=== Evaluating {test_year} ===")
         year_result = evaluate_year(test_year, df, review_vectorizer, clf)
         results.append(year_result)
@@ -106,15 +119,21 @@ def main():
     print("\n=== Final Results ===")
     print(f"Correctly predicted winners in {correct_winners} of {len(results_df)} years")
 
-    # Plot actual winners' predicted probabilities
     plt.figure(figsize=(10, 6))
-    plt.plot(results_df['year'], results_df['actual_prob'], marker='o', linestyle='-', color='blue')
-    plt.title('Predicted Probability for Actual Winners Over the Years')
-    plt.xlabel('Year')
-    plt.ylabel('Normalized Predicted Probability')
+    plt.plot(
+        results_df["year"],
+        results_df["actual_prob"],
+        marker="o",
+        linestyle="-",
+        color="blue",
+    )
+    plt.title("Predicted Probability for Actual Winners Over the Years")
+    plt.xlabel("Year")
+    plt.ylabel("Normalized Predicted Probability")
     plt.ylim(0, 1)
     plt.grid(True)
     plt.show()
+
 
 if __name__ == "__main__":
     main()
